Decouple the Contract from the compiler module
