+++ src/compiler/mod.rs
@@ -11,14 +11,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use crate::contract::data_table::DataTable;
+use crate::contract::Contract;
 use crate::interpreter::types::{Numeric, PactType, StringLike};
 use crate::interpreter::OpCode;
 use crate::parser::ast;
 
-mod data_table;
-pub use data_table::DataTable;
-
-use bit_reverse::ParallelReverse;
 use hashbrown::HashMap;
 
 /// Compilation error
@@ -28,51 +26,6 @@ pub enum CompileErr {
     UndeclaredVar(ast::Identifier),
     /// A parameter with the same identifier has already been declared
     Redeclared,
-}
-
-#[derive(Debug, PartialEq)]
-/// A binary format error
-pub enum BinaryFormatErr {
-    /// Version mismatch
-    UnsupportedVersion,
-    /// DataTable is invalid
-    MalformedDataTable(&'static str),
-    // The buffer is to short to be valid
-    TooShort,
-}
-
-/// A pact contract
-/// It has byte code and an accompanying data section
-#[derive(Debug, PartialEq)]
-pub struct Contract<'a> {
-    pub data_table: DataTable<'a>,
-    pub bytecode: Vec<u8>,
-}
-
-impl<'a> Contract<'a> {
-    /// Encode the contract as v0 binary format into `buf`
-    pub fn encode(&self, buf: &mut Vec<u8>) {
-        buf.push(0); // binary format version: `0`
-        self.data_table.encode(buf);
-        buf.extend(self.bytecode.clone());
-    }
-    /// Decode a pact contract from v0 binary format
-    pub fn decode(buf: &'a [u8]) -> Result<Self, BinaryFormatErr> {
-        if buf.len() < 2 {
-            println!("len: {:?}", buf.len());
-            return Err(BinaryFormatErr::TooShort);
-        }
-        if buf[0].swap_bits() != 0 {
-            return Err(BinaryFormatErr::UnsupportedVersion);
-        }
-        let (data_table, offset) =
-            DataTable::decode(&buf[1..]).map_err(|err| BinaryFormatErr::MalformedDataTable(err))?;
-        let bytecode = buf[1usize + offset..].to_vec();
-        Ok(Self {
-            data_table,
-            bytecode,
-        })
-    }
 }
 
 /// Compile a pact contract AST into bytecode
@@ -228,22 +181,4 @@ fn compile_comparator(comparator: &ast::Comparator) -> Result<OpCode, CompileErr
         ast::Comparator::LessThanOrEqual => OpCode::LTE,
     }
     .into())
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn contract_binary_format_unsupported_version() {
-        assert_eq!(
-            Contract::decode(&[1, 0]),
-            Err(BinaryFormatErr::UnsupportedVersion)
-        );
-    }
-
-    #[test]
-    fn contract_binary_format_too_short() {
-        assert_eq!(Contract::decode(&[0]), Err(BinaryFormatErr::TooShort));
-    }
 }
